[upstream] Sensor clean up (box2d #953)
- Removed b2BodyDef::enableSensorHits
- Added access to joint event thresholds
- Removed b2SensorData
- motion locks no longer affect mass
- all visitors now do continuous checks against static sensors
- bullets visitors do continuous checks against all non-bullet sensors
- gave up on supporting fast moving sensors or collecting sensor
collision data
- instead people should use ray and shape casts to get detailed swept
collision data or use contact events
- added sample that shows how to make an exploding projectile

+++ src/Box2D.NET/B2SolverSets.cs
@@ -76,7 +76,7 @@ namespace Box2D.NET
                 ref B2BodyState state = ref b2Array_Add(ref awakeSet.bodyStates);
                 //*state = b2_identityBodyState;
                 state.CopyFrom(b2_identityBodyState);
-                state.flags = (int)body.flags;
+                state.flags = body.flags;
 
                 // move non-touching contacts from disabled set to awake set
                 int contactKey = body.headContactKey;
@@ -580,7 +580,7 @@ namespace Box2D.NET
                 ref B2BodyState state = ref b2Array_Add(ref targetSet.bodyStates);
                 //*state = b2_identityBodyState;
                 state.CopyFrom(b2_identityBodyState);
-                state.flags = (int)body.flags;
+                state.flags = body.flags;
             }
 
             body.setIndex = targetSet.setIndex;
